moved integration module into reconcile/change_owner directory
this is the last step of moving files around :)
i also made sure all the modules are named consistently camelcasing words
